// projects/dynamic-forms/src/lib/angular-material/models/dynamic-form-control.model.ts
export interface DynamicFormControlCustomEvent {

  type: string;                    // 'click' string
  id: string;                      // 'organisation.name'
  directive: string;               // 'matSuffix'
  name: string;                    // 'search'

}

export interface DynamicFormControl {

  formGroup: FormGroup;
  model: DynamicFormControlModel;

  customEvent?: EventEmitter<any>;

}

export interface DynamicSelect{
  name: string;
  value: string;
}

export interface DynamicFormControlModelConfig {

  // Mandatory items

  type: string;                    // "input"
  id: string;                      // "givenName"

  inputType?: string;              // "text"

  // Optional items

  appearance?: string;             // "outline"
  autocomplete?: string;           // aka autoFill
  gridItemClass?: string;          // "grid-column-1"
  hideRequiredMarker?: boolean;    // false
  label?: string;                  // "Given Name"
  name?: string;                   // "givenName"
  placeholder?: string;            // "Given Name"
  prefixIconName?: string;
  required?: boolean;              // false
  suffixIconName?: string;

  validators?: ValidatorModel[];

  //-- my new properties

  source?: string;

  checked?: boolean;

  dataSet?: DynamicSelect[];

}

export class DynamicFormControlModel {

  // Mandatory items

  type: string;                    // "input"
  id: string;                      // "givenName"

  inputType: string;               // "text"

  // Optional items

  appearance?: string;             // "outline"
  autocomplete?: string;           // aka autoFill
  gridItemClass?: string;          // "grid-column-1"
  hideRequiredMarker?: boolean;    // false
  label?: string;                  // "Given Name"
  name?: string;                   // "givenName"
  placeholder?: string;            // "Given Name"
  prefixIconName?: string;
  required?: boolean;              // false
  suffixIconName?: string;

  validators?: ValidatorModel[];

  source?: string; // if media source
  
  checked?: boolean;

  dataSet?: DynamicSelect[];



  public constructor(config: DynamicFormControlModelConfig) {

    this.type = config.type;
    this.id = config.id;

    this.inputType = config.inputType || 'text';

    this.appearance = config.appearance || 'outline';
    this.autocomplete = config.autocomplete || null;
    this.gridItemClass = config.gridItemClass || null;
    this.hideRequiredMarker = config.hideRequiredMarker || false;
    this.label = config.label || null;
    this.name = config.name || config.id;
    this.placeholder = config.placeholder || config.label;
    this.prefixIconName = config.prefixIconName || null;
    this.required = config.required || false;
    this.suffixIconName = config.suffixIconName || null;

    this.validators = config.validators || null;

    this.source = config.source || null;

    this.checked = config.checked || false;

    this.dataSet = config.dataSet || [];

  }

}
// ...
export interface DynamicFormControlModel {

  // Container i.e., mat-form-field

  appearance?: string;
  class?: string;

  // Label i.e., mat-label

  label?: string;

  // Control i.e., input, select, ...

  type?: string;          // 'input'
  autocomplete?: string;  // aka autoFill
  id?: string;
  name?: string;          // name || id
  placeholder?: string;
  required?: boolean;
  hideRequiredMarker?: boolean;

  validators?: ValidatorModel[];

  // config {}            // aka attributes ???
  // options {}

}